Added building placement and tree placement to terrain generator.
Generator can now place buildings on the surface of the cube via the PlaceBuilding() function.

Terrain generation now covers a specified percentage of the map with trees at start.

Played around with some fractal tree generation but it didn't look very good so I'll stick with the "ball and post" style tree for now.